requirements.txt
streamlit
torch
torchvision
pillow
streamlit-webrtc
opencv-python-headless
av